src/requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
PyYAML==6.0.1
